+++ CHANGELOG.md
@@ -75,8 +75,11 @@
 * `desktop-notifications.command` option, replaces `notify`.
 * `desktop-notifications.inhibit-when-focused` option, replaces
   `notify-focus-inhibit`.
-* `${icon}`, `${urgency}`,`${action-name}` and `${action-label}` added
+* `${icon}`, `${urgency}` and `${action-argument}` added
   to the `desktop-notifications.command` template.
+* `desktop-notifications.command-action-argument` option, defining how
+  `${action-argument}` (in `desktop-notifications.command`) should be
+  expanded.
 * `desktop-notifications.close` option, defining what to execute when
   an application wants to close an existing notification (via an
   OSC-99 escape sequence).